Move TPH inheritance tests into a dedicated suite
Extract the single-table-inheritance tests and their models from
InheritanceTests into a new TphInheritanceTests fixture and drop the
issue-number framing now that they cover general TPH behavior.

Add projection-downcast edge cases: whole-row (T)x, downcast-then-member
((T)x).Member on filtered and mixed rows, and member-then-cast (T)x.Assoc.
